+++ lib/ddp_ejson.dart
@@ -33,4 +33,3 @@ class Doc {
     return item;
   }
 }
-